pyproject.toml
[tool.poetry]
name = "music streaming app - mad-1 project"
version = "1.0.0"
description = "A project for the MAD-1 course in the IITM B.Sc Degree"
authors = ["Vishal N <dev@vishalnandagopal.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.0"
flask-session = "^0.5.0"
python-dotenv = "^1.0.0"
faker = "^21.0.0"


[tool.poetry.group.dev.dependencies]
icecream = "^2.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
